+++ lore-integration-tests/tests/store_keep_alive_test.rs
@@ -225,6 +225,7 @@ mod store_keep_alive_tests {
                 no_commit: 0,
                 link: Default::default(),
                 ignore_links: 0,
+                inherit_metadata: Default::default(),
             },
             None,
         )