+++ lore-revision/src/file/history.rs
@@ -372,6 +372,12 @@ async fn find_start_revision(
         resolved.id
     };
 
+    if execution_context().globals().offline_or_local() {
+        return Ok(branch::load_latest(repository.clone(), branch)
+            .await
+            .unwrap_or_default());
+    }
+
     let remote_latest = if let Ok(remote) = repository.remote().await {
         branch::load_remote_latest(remote.clone(), repository.id, branch)
             .await
@@ -386,9 +392,6 @@ async fn find_start_revision(
     let local_latest = branch::load_latest(repository.clone(), branch)
         .await
         .unwrap_or_default();
-    if execution_context().globals().local() {
-        return Ok(local_latest);
-    }
 
     // Is there a latest?
     if remote_latest.is_zero() {