+++ lore-revision/src/stage.rs
@@ -3003,9 +3003,11 @@ pub(crate) async fn stage_from_parent_revision(
         .serialize(repository.clone(), token)
         .await
         .forward::<StageError>("Failed to serialize staged revision state")?;
-    crate::instance::store_staged_anchor(&repository, signature)
-        .await
-        .forward::<StageError>("Failed to serialize staged anchor")?;
+    if !execution_context().globals().dry_run() {
+        crate::instance::store_staged_anchor(&repository, signature)
+            .await
+            .forward::<StageError>("Failed to serialize staged anchor")?;
+    }
 
     event::LoreEvent::FileStageRevision(LoreFileStageRevisionEventData {
         repository: repository.id,
@@ -3366,9 +3368,11 @@ pub(crate) async fn stage_link_paths_from_parent_revision(
         .serialize(repository.clone(), token)
         .await
         .forward::<StageError>("Failed to serialize parent staged state")?;
-    crate::instance::store_staged_anchor(&repository, signature)
-        .await
-        .forward::<StageError>("Failed to store parent staged anchor")?;
+    if !execution_context().globals().dry_run() {
+        crate::instance::store_staged_anchor(&repository, signature)
+            .await
+            .forward::<StageError>("Failed to store parent staged anchor")?;
+    }
 
     Ok(())
 }